bf RDM-782 исправление типа системного идентификатора записи.

+++ rdm-n2o-l10n/src/main/java/ru/i_novus/ms/rdm/n2o/l10n/resolver/L10nLocalizeRecordGetterResolver.java
@@ -89,7 +89,7 @@ public class L10nLocalizeRecordGetterResolver implements DataRecordGetterResolve
 
         SearchDataCriteria dataCriteria = new SearchDataCriteria();
         dataCriteria.setLocaleCode(criteria.getLocaleCode());
-        dataCriteria.setRowSystemIds(singletonList(criteria.getId().longValue()));
+        dataCriteria.setRowSystemIds(singletonList(criteria.getId()));
 
         Page<RefBookRowValue> rowValues = versionService.search(criteria.getVersionId(), dataCriteria);
         return !isEmpty(rowValues.getContent()) ? rowValues.getContent() : emptyList();
